requirements.txt
python-dotenv==0.20.0
bcrypt~=3.2.0
SQLAlchemy~=1.4.34